requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
flask
werkzeug
mlflow
joblib
PyYAML
optuna